move invocation_id from events to first class common concept
